requirements.txt
fastapi
uvicorn[standard]
gunicorn
google-genai
python-dotenv
pydantic
